First stab at DirectoryHandleMount
