+++ packages/php-wasm/util/src/lib/index.ts
@@ -2,7 +2,13 @@ import Semaphore, { AcquireTimeoutError } from './semaphore';
 export { Semaphore, AcquireTimeoutError };
 export { PhpWasmError } from './php-wasm-error';
 export type { SemaphoreOptions } from './semaphore';
-export { dirname, joinPaths, basename, normalizePath } from './paths';
+export {
+	dirname,
+	joinPaths,
+	basename,
+	normalizePath,
+	isParentOf,
+} from './paths';
 export { createSpawnHandler } from './create-spawn-handler';
 export { randomString } from './random-string';
 export { randomFilename } from './random-filename';
