+++ packages/php-wasm/web/src/lib/directory-handle-mount.ts
@@ -1,0 +1,410 @@
+import {
+	Emscripten,
+	FSHelpers,
+	Mountable,
+	PHP,
+	__private__dont__use,
+} from '@php-wasm/universal';
+import { Semaphore, isParentOf, joinPaths } from '@php-wasm/util';
+import { logger } from '@php-wasm/logger';
+import { FilesystemOperation, journalFSEvents } from '@php-wasm/fs-journal';
+
+export interface MountOptions {
+	initialSync: {
+		direction?: 'opfs-to-memfs' | 'memfs-to-opfs';
+		onProgress?: SyncProgressCallback;
+	};
+}
+export type SyncProgress = {
+	/** The number of files that have been synced. */
+	files: number;
+	/** The number of all files that need to be synced. */
+	total: number;
+};
+export type SyncProgressCallback = (progress: SyncProgress) => void;
+
+const OPFSMounts = new WeakMap<PHP, Record<string, DirectoryHandleMount>>();
+
+export class DirectoryHandleMount implements Mountable {
+	private options: MountOptions;
+	private mountPoint: string | undefined;
+	private unbindJournal: () => any = () => {};
+	constructor(
+		private readonly php: PHP,
+		private readonly handle: FileSystemDirectoryHandle,
+		options: MountOptions = { initialSync: {} }
+	) {
+		this.options = {
+			...options,
+			initialSync: {
+				...options.initialSync,
+				direction: options.initialSync.direction ?? 'opfs-to-memfs',
+			},
+		};
+	}
+
+	async mount(FS: Emscripten.RootFS, vfsMountPoint: string): Promise<void> {
+		if (!OPFSMounts.has(this.php)) {
+			OPFSMounts.set(this.php, {});
+		}
+		const existingMounts = OPFSMounts.get(this.php)!;
+
+		// Ensure there are no other conflicting mounts.
+		if (vfsMountPoint in existingMounts) {
+			for (const path in Object.values(existingMounts)) {
+				if (
+					isParentOf(path, vfsMountPoint) ||
+					isParentOf(vfsMountPoint, path)
+				) {
+					throw new Error(
+						`Cannot mount ${vfsMountPoint} because it conflicts with a previously mounted ${path}`
+					);
+				}
+			}
+		}
+
+		if (this.mountPoint) {
+			throw new Error('Already mounted');
+		}
+
+		this.mountPoint = vfsMountPoint;
+
+		if (this.options.initialSync.direction === 'opfs-to-memfs') {
+			if (FSHelpers.fileExists(FS, vfsMountPoint)) {
+				FSHelpers.rmdir(FS, vfsMountPoint);
+			}
+			FSHelpers.mkdir(FS, vfsMountPoint);
+			await copyOpfsToMemfs(FS, this.handle, vfsMountPoint);
+		} else {
+			await copyMemfsToOpfs(
+				FS,
+				this.handle,
+				vfsMountPoint,
+				this.options.initialSync.onProgress
+			);
+		}
+		this.unbindJournal = journalFSEventsToOpfs(
+			this.php,
+			this.handle,
+			vfsMountPoint
+		);
+		existingMounts[vfsMountPoint] = this;
+	}
+
+	unmount() {
+		if (this.mountPoint) {
+			throw new Error('Not mounted yet');
+		}
+		this.unbindJournal();
+		const existingMounts = OPFSMounts.get(this.php)!;
+		delete existingMounts[this.mountPoint!];
+		delete this.mountPoint;
+	}
+}
+
+async function copyOpfsToMemfs(
+	FS: Emscripten.RootFS,
+	opfsRoot: FileSystemDirectoryHandle,
+	memfsRoot: string
+) {
+	FSHelpers.mkdir(FS, memfsRoot);
+
+	/**
+	 * Semaphores are used to limit the number of concurrent operations.
+	 * Flooding the browser with 2000 FS operations at the same time
+	 * can get quite slow.
+	 */
+	const semaphore = new Semaphore({
+		concurrency: 40,
+	});
+
+	const ops: Array<Promise<void>> = [];
+	const stack: Array<[FileSystemDirectoryHandle, string]> = [
+		[opfsRoot, memfsRoot],
+	];
+	while (stack.length > 0) {
+		const [opfsParent, memfsParentPath] = stack.pop()!;
+
+		for await (const opfsHandle of opfsParent.values()) {
+			const op = semaphore.run(async () => {
+				const memfsEntryPath = joinPaths(
+					memfsParentPath,
+					opfsHandle.name
+				);
+				if (opfsHandle.kind === 'directory') {
+					try {
+						FS.mkdir(memfsEntryPath);
+					} catch (e) {
+						if ((e as any)?.errno !== 20) {
+							logger.error(e);
+							// We ignore the error if the directory already exists,
+							// and throw otherwise.
+							throw e;
+						}
+					}
+					stack.push([opfsHandle, memfsEntryPath]);
+				} else if (opfsHandle.kind === 'file') {
+					const file = await opfsHandle.getFile();
+					const byteArray = new Uint8Array(await file.arrayBuffer());
+					FS.createDataFile(
+						memfsParentPath,
+						opfsHandle.name,
+						byteArray,
+						true,
+						true,
+						true
+					);
+				}
+				ops.splice(ops.indexOf(op), 1);
+			});
+			ops.push(op);
+		}
+		// Let the ongoing operations catch-up to the stack.
+		while (stack.length === 0 && ops.length > 0) {
+			await Promise.any(ops);
+		}
+	}
+}
+
+export async function copyMemfsToOpfs(
+	FS: Emscripten.RootFS,
+	opfsRoot: FileSystemDirectoryHandle,
+	memfsRoot: string,
+	onProgress?: SyncProgressCallback
+) {
+	// Ensure the memfs directory exists.
+	FS.mkdirTree(memfsRoot);
+
+	// Create all MEMFS directories in OPFS but don't create
+	// files yet. This is quite fast.
+	const filesToCreate: Array<[FileSystemDirectoryHandle, string, string]> =
+		[];
+	async function mirrorMemfsDirectoryinOpfs(
+		memfsParent: string,
+		opfsDir: FileSystemDirectoryHandle
+	) {
+		await Promise.all(
+			FS.readdir(memfsParent)
+				.filter(
+					(entryName: string) =>
+						entryName !== '.' && entryName !== '..'
+				)
+				.map(async (entryName: string) => {
+					const memfsPath = joinPaths(memfsParent, entryName);
+					if (!isMemfsDir(FS, memfsPath)) {
+						filesToCreate.push([opfsDir, memfsPath, entryName]);
+						return;
+					}
+
+					const handle = await opfsDir.getDirectoryHandle(entryName, {
+						create: true,
+					});
+					return await mirrorMemfsDirectoryinOpfs(memfsPath, handle);
+				})
+		);
+	}
+	await mirrorMemfsDirectoryinOpfs(memfsRoot, opfsRoot);
+
+	// Now let's create all the required files in OPFS. This is quite slow
+	// so we report progress.
+	let i = 0;
+	const filesCreated = filesToCreate.map(([opfsDir, memfsPath, entryName]) =>
+		overwriteOpfsFile(opfsDir, entryName, FS, memfsPath).then(() => {
+			onProgress?.({ files: ++i, total: filesToCreate.length });
+		})
+	);
+	await Promise.all(filesCreated);
+}
+
+function isMemfsDir(FS: Emscripten.RootFS, path: string) {
+	return FS.isDir(FS.lookupPath(path, { follow: true }).node.mode);
+}
+
+async function overwriteOpfsFile(
+	opfsParent: FileSystemDirectoryHandle,
+	name: string,
+	FS: Emscripten.RootFS,
+	memfsPath: string
+) {
+	let buffer;
+	try {
+		buffer = FS.readFile(memfsPath, {
+			encoding: 'binary',
+		});
+	} catch (e) {
+		// File was removed, ignore
+		return;
+	}
+
+	const opfsFile = await opfsParent.getFileHandle(name, { create: true });
+	const writer =
+		opfsFile.createWritable !== undefined
+			? // Google Chrome, Firefox, probably more browsers
+			  await opfsFile.createWritable()
+			: // Safari
+			  await opfsFile.createSyncAccessHandle();
+	try {
+		await writer.truncate(0);
+		await writer.write(buffer);
+	} finally {
+		await writer.close();
+	}
+}
+
+export function journalFSEventsToOpfs(
+	php: PHP,
+	opfsRoot: FileSystemDirectoryHandle,
+	memfsRoot: string
+) {
+	const journal: FilesystemOperation[] = [];
+	const unbindJournal = journalFSEvents(php, memfsRoot, (entry) => {
+		journal.push(entry);
+	});
+	const rewriter = new OpfsRewriter(php, opfsRoot, memfsRoot);
+
+	async function flushJournal() {
+		const release = await php.semaphore.acquire();
+		try {
+			// @TODO This is way too slow in practice, we need to batch the
+			// changes into groups of parallelizable operations.
+			while (journal.length) {
+				await rewriter.processEntry(journal.shift()!);
+			}
+		} finally {
+			release();
+		}
+	}
+	php.addEventListener('request.end', flushJournal);
+	return function () {
+		unbindJournal();
+		php.removeEventListener('request.end', flushJournal);
+	};
+}
+
+type JournalEntry = FilesystemOperation;
+
+class OpfsRewriter {
+	private memfsRoot: string;
+
+	constructor(
+		private php: PHP,
+		private opfs: FileSystemDirectoryHandle,
+		memfsRoot: string
+	) {
+		this.memfsRoot = normalizeMemfsPath(memfsRoot);
+	}
+
+	private toOpfsPath(path: string) {
+		return normalizeMemfsPath(path.substring(this.memfsRoot.length));
+	}
+
+	public async processEntry(entry: JournalEntry) {
+		if (
+			!entry.path.startsWith(this.memfsRoot) ||
+			entry.path === this.memfsRoot
+		) {
+			return;
+		}
+		const opfsPath = this.toOpfsPath(entry.path);
+		const opfsParent = await resolveParent(this.opfs, opfsPath);
+		const name = getFilename(opfsPath);
+		if (!name) {
+			return;
+		}
+
+		try {
+			if (entry.operation === 'DELETE') {
+				try {
+					await opfsParent.removeEntry(name, {
+						recursive: true,
+					});
+				} catch (e) {
+					// If the directory already doesn't exist, it's fine
+				}
+			} else if (entry.operation === 'CREATE') {
+				if (entry.nodeType === 'directory') {
+					await opfsParent.getDirectoryHandle(name, {
+						create: true,
+					});
+				} else {
+					await opfsParent.getFileHandle(name, {
+						create: true,
+					});
+				}
+			} else if (entry.operation === 'WRITE') {
+				await overwriteOpfsFile(
+					opfsParent,
+					name,
+					this.php[__private__dont__use].FS,
+					entry.path
+				);
+			} else if (
+				entry.operation === 'RENAME' &&
+				entry.toPath.startsWith(this.memfsRoot)
+			) {
+				const opfsTargetPath = this.toOpfsPath(entry.toPath);
+				const opfsTargetParent = await resolveParent(
+					this.opfs,
+					opfsTargetPath
+				);
+				const targetName = getFilename(opfsTargetPath);
+
+				if (entry.nodeType === 'directory') {
+					const opfsDir = await opfsTargetParent.getDirectoryHandle(
+						name,
+						{
+							create: true,
+						}
+					);
+					// in OPFS, move() doesn't work for directories :-(
+					// We have to copy the directory recursively instead.
+					await copyMemfsToOpfs(
+						this.php[__private__dont__use].FS,
+						opfsDir,
+						entry.toPath
+					);
+					// Then delete the old directory
+					await opfsParent.removeEntry(name, {
+						recursive: true,
+					});
+				} else {
+					const file = await opfsParent.getFileHandle(name);
+					file.move(opfsTargetParent, targetName);
+				}
+			}
+		} catch (e) {
+			// Useful for debugging – the original error gets lost in the
+			// Comlink proxy.
+			logger.log({ entry, name });
+			logger.error(e);
+			throw e;
+		}
+	}
+}
+
+function normalizeMemfsPath(path: string) {
+	return path.replace(/\/$/, '').replace(/\/\/+/g, '/');
+}
+
+function getFilename(path: string) {
+	return path.substring(path.lastIndexOf('/') + 1);
+}
+
+async function resolveParent(
+	opfs: FileSystemDirectoryHandle,
+	relativePath: string
+): Promise<FileSystemDirectoryHandle> {
+	const normalizedPath = relativePath
+		.replace(/^\/+|\/+$/g, '')
+		.replace(/\/+/, '/');
+	if (!normalizedPath) {
+		return opfs;
+	}
+	const segments = normalizedPath.split('/');
+	let handle: FileSystemDirectoryHandle | FileSystemFileHandle = opfs;
+	for (let i = 0; i < segments.length - 1; i++) {
+		const segment = segments[i];
+		handle = await handle.getDirectoryHandle(segment, { create: true });
+	}
+	return handle as any;
+}
