+++ packages/php-wasm/web/src/lib/index.ts
@@ -7,3 +7,9 @@ export { getPHPLoaderModule } from './get-php-loader-module';
 export { registerServiceWorker } from './register-service-worker';
 
 export { spawnPHPWorkerThread } from './worker-thread/spawn-php-worker-thread';
+export { DirectoryHandleMount } from './directory-handle-mount';
+export type {
+	MountOptions,
+	SyncProgress,
+	SyncProgressCallback,
+} from './directory-handle-mount';
